+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 2.6.0 (Unreleased)
 
+### Changed
+#### SidebarBundle
+* Class `Oro\Bundle\SidebarBundle\Model\WidgetDefinitionRegistry`
+    * the return type for methods `getWidgetDefinitions` and `getWidgetDefinitionsByPlacement` were changed from `ArrayCollection` to `array`
 ### Removed
 #### AttachmentBundle
 * The parameter `oro_attachment.listener.file_listener.class` was removed form the service container
@@ -9,13 +13,17 @@
 * The parameter `oro_imap.listener.user_email_origin.class` was removed form the service container
 #### ReminderBundle
 * The parameter `oro_reminder.event_listener.reminder_listener.class` was removed form the service container
+#### SidebarBundle
+* The parameter `oro_sidebar.widget_definition.registry.class` was removed form the service container
+* The service `oro_sidebar.request.handler` was removed
 #### SSOBundle
 * The parameter `oro_sso.event_listener.user_email_change_listener.class` was removed form the service container
 #### WorkflowBundle
 * The parameter `oro_workflow.listener.process_data_serialize.class` was removed form the service container
 * The parameter `oro_workflow.listener.workflow_data_serialize.class` was removed form the service container
 
-## 2.5.0 (Unreleased)
+## 2.5.0 (2017-11-30)
+[Show detailed list of changes](incompatibilities-2-5.md)
 
 ### Added
 #### ActivityListBundle